Fix code smells found by SonarCloud (#261)
* Fix bug found by SonarCloud

* Fix critical code smells found by SonarCloud

* Fix major code smells found by SonarCloud

* Fix minor code smells found by SonarCloud

* Update ruuvi.json_stream_gen.c to the master branch

+++ src/http_server_auth_digest.c
@@ -25,14 +25,14 @@ http_server_parse_token(
         return false;
     }
     p_idx1 += strlen(p_prefix);
-    const ptrdiff_t   ptr_diff = (ptrdiff_t)(p_idx1 - p_str);
+    const ptrdiff_t   ptr_diff = p_idx1 - p_str;
     const size_t      len_diff = len - ptr_diff;
     const char* const p_idx2   = http_server_strnstr(p_idx1, p_suffix, len_diff);
     if (NULL == p_idx2)
     {
         return false;
     }
-    const size_t token_len = (size_t)(ptrdiff_t)(p_idx2 - p_idx1);
+    const size_t token_len = (size_t)(p_idx2 - p_idx1);
     if (token_len >= buf_size)
     {
         return false;
